# back-end/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: spaghetti-back-end
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /createUser/{userName}/password/{password}:
    # binds a127 app logic to a route
    x-swagger-router-controller: user_ops
    post:
      description: Create a new user
      # used as the method name of the controller
      operationId: createUser
      parameters:
        - name: userName
          in: path
          description: The desired username
          required: true
          type: string
        - name: password
          in: path
          description: The desired password
          required: true
          type: string
      responses:
        "201":
          description: User creation success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/genericResponse"
        "400":
          description: User name already exists
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /retrieveUser/{userName}/password/{password}:
    # binds a127 app logic to a route
    x-swagger-router-controller: user_ops
    get:
      description: Retrieve an existing user
      # used as the method name of the controller
      operationId: retrieveUser
      parameters:
        - name: userName
          in: path
          description: The existing username
          required: true
          type: string
        - name: password
          in: path
          description: The existing password
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/genericResponse"
        "400":
          description: User name does not exist
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Password incorrect
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  genericResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string